Share stimulus test MAT fixtures
